+++ bccsp/handlers/mock/credrequest.go
@@ -4,19 +4,17 @@ package mock
 import (
 	"sync"
 
-	"github.com/IBM/idemix/bccsp/schemes/dlog/handlers"
+	"github.com/IBM/idemix/bccsp/handlers"
 	math "github.com/IBM/mathlib"
 )
 
-type NymSignatureScheme struct {
-	SignStub        func(*math.Zr, *math.G1, *math.Zr, handlers.IssuerPublicKey, []byte) ([]byte, error)
+type CredRequest struct {
+	SignStub        func(*math.Zr, handlers.IssuerPublicKey, []byte) ([]byte, error)
 	signMutex       sync.RWMutex
 	signArgsForCall []struct {
 		arg1 *math.Zr
-		arg2 *math.G1
-		arg3 *math.Zr
-		arg4 handlers.IssuerPublicKey
-		arg5 []byte
+		arg2 handlers.IssuerPublicKey
+		arg3 []byte
 	}
 	signReturns struct {
 		result1 []byte
@@ -26,13 +24,12 @@ type NymSignatureScheme struct {
 		result1 []byte
 		result2 error
 	}
-	VerifyStub        func(handlers.IssuerPublicKey, *math.G1, []byte, []byte) error
+	VerifyStub        func([]byte, handlers.IssuerPublicKey, []byte) error
 	verifyMutex       sync.RWMutex
 	verifyArgsForCall []struct {
-		arg1 handlers.IssuerPublicKey
-		arg2 *math.G1
+		arg1 []byte
+		arg2 handlers.IssuerPublicKey
 		arg3 []byte
-		arg4 []byte
 	}
 	verifyReturns struct {
 		result1 error
@@ -44,27 +41,25 @@ type NymSignatureScheme struct {
 	invocationsMutex sync.RWMutex
 }
 
-func (fake *NymSignatureScheme) Sign(arg1 *math.Zr, arg2 *math.G1, arg3 *math.Zr, arg4 handlers.IssuerPublicKey, arg5 []byte) ([]byte, error) {
-	var arg5Copy []byte
-	if arg5 != nil {
-		arg5Copy = make([]byte, len(arg5))
-		copy(arg5Copy, arg5)
+func (fake *CredRequest) Sign(arg1 *math.Zr, arg2 handlers.IssuerPublicKey, arg3 []byte) ([]byte, error) {
+	var arg3Copy []byte
+	if arg3 != nil {
+		arg3Copy = make([]byte, len(arg3))
+		copy(arg3Copy, arg3)
 	}
 	fake.signMutex.Lock()
 	ret, specificReturn := fake.signReturnsOnCall[len(fake.signArgsForCall)]
 	fake.signArgsForCall = append(fake.signArgsForCall, struct {
 		arg1 *math.Zr
-		arg2 *math.G1
-		arg3 *math.Zr
-		arg4 handlers.IssuerPublicKey
-		arg5 []byte
-	}{arg1, arg2, arg3, arg4, arg5Copy})
+		arg2 handlers.IssuerPublicKey
+		arg3 []byte
+	}{arg1, arg2, arg3Copy})
 	stub := fake.SignStub
 	fakeReturns := fake.signReturns
-	fake.recordInvocation("Sign", []interface{}{arg1, arg2, arg3, arg4, arg5Copy})
+	fake.recordInvocation("Sign", []interface{}{arg1, arg2, arg3Copy})
 	fake.signMutex.Unlock()
 	if stub != nil {
-		return stub(arg1, arg2, arg3, arg4, arg5)
+		return stub(arg1, arg2, arg3)
 	}
 	if specificReturn {
 		return ret.result1, ret.result2
@@ -72,26 +67,26 @@ func (fake *NymSignatureScheme) Sign(arg1 *math.Zr, arg2 *math.G1, arg3 *math.Zr
 	return fakeReturns.result1, fakeReturns.result2
 }
 
-func (fake *NymSignatureScheme) SignCallCount() int {
+func (fake *CredRequest) SignCallCount() int {
 	fake.signMutex.RLock()
 	defer fake.signMutex.RUnlock()
 	return len(fake.signArgsForCall)
 }
 
-func (fake *NymSignatureScheme) SignCalls(stub func(*math.Zr, *math.G1, *math.Zr, handlers.IssuerPublicKey, []byte) ([]byte, error)) {
+func (fake *CredRequest) SignCalls(stub func(*math.Zr, handlers.IssuerPublicKey, []byte) ([]byte, error)) {
 	fake.signMutex.Lock()
 	defer fake.signMutex.Unlock()
 	fake.SignStub = stub
 }
 
-func (fake *NymSignatureScheme) SignArgsForCall(i int) (*math.Zr, *math.G1, *math.Zr, handlers.IssuerPublicKey, []byte) {
+func (fake *CredRequest) SignArgsForCall(i int) (*math.Zr, handlers.IssuerPublicKey, []byte) {
 	fake.signMutex.RLock()
 	defer fake.signMutex.RUnlock()
 	argsForCall := fake.signArgsForCall[i]
-	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3, argsForCall.arg4, argsForCall.arg5
+	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3
 }
 
-func (fake *NymSignatureScheme) SignReturns(result1 []byte, result2 error) {
+func (fake *CredRequest) SignReturns(result1 []byte, result2 error) {
 	fake.signMutex.Lock()
 	defer fake.signMutex.Unlock()
 	fake.SignStub = nil
@@ -101,7 +96,7 @@ func (fake *NymSignatureScheme) SignReturns(result1 []byte, result2 error) {
 	}{result1, result2}
 }
 
-func (fake *NymSignatureScheme) SignReturnsOnCall(i int, result1 []byte, result2 error) {
+func (fake *CredRequest) SignReturnsOnCall(i int, result1 []byte, result2 error) {
 	fake.signMutex.Lock()
 	defer fake.signMutex.Unlock()
 	fake.SignStub = nil
@@ -117,31 +112,30 @@ func (fake *NymSignatureScheme) SignReturnsOnCall(i int, result1 []byte, result2
 	}{result1, result2}
 }
 
-func (fake *NymSignatureScheme) Verify(arg1 handlers.IssuerPublicKey, arg2 *math.G1, arg3 []byte, arg4 []byte) error {
+func (fake *CredRequest) Verify(arg1 []byte, arg2 handlers.IssuerPublicKey, arg3 []byte) error {
+	var arg1Copy []byte
+	if arg1 != nil {
+		arg1Copy = make([]byte, len(arg1))
+		copy(arg1Copy, arg1)
+	}
 	var arg3Copy []byte
 	if arg3 != nil {
 		arg3Copy = make([]byte, len(arg3))
 		copy(arg3Copy, arg3)
 	}
-	var arg4Copy []byte
-	if arg4 != nil {
-		arg4Copy = make([]byte, len(arg4))
-		copy(arg4Copy, arg4)
-	}
 	fake.verifyMutex.Lock()
 	ret, specificReturn := fake.verifyReturnsOnCall[len(fake.verifyArgsForCall)]
 	fake.verifyArgsForCall = append(fake.verifyArgsForCall, struct {
-		arg1 handlers.IssuerPublicKey
-		arg2 *math.G1
+		arg1 []byte
+		arg2 handlers.IssuerPublicKey
 		arg3 []byte
-		arg4 []byte
-	}{arg1, arg2, arg3Copy, arg4Copy})
+	}{arg1Copy, arg2, arg3Copy})
 	stub := fake.VerifyStub
 	fakeReturns := fake.verifyReturns
-	fake.recordInvocation("Verify", []interface{}{arg1, arg2, arg3Copy, arg4Copy})
+	fake.recordInvocation("Verify", []interface{}{arg1Copy, arg2, arg3Copy})
 	fake.verifyMutex.Unlock()
 	if stub != nil {
-		return stub(arg1, arg2, arg3, arg4)
+		return stub(arg1, arg2, arg3)
 	}
 	if specificReturn {
 		return ret.result1
@@ -149,26 +143,26 @@ func (fake *NymSignatureScheme) Verify(arg1 handlers.IssuerPublicKey, arg2 *math
 	return fakeReturns.result1
 }
 
-func (fake *NymSignatureScheme) VerifyCallCount() int {
+func (fake *CredRequest) VerifyCallCount() int {
 	fake.verifyMutex.RLock()
 	defer fake.verifyMutex.RUnlock()
 	return len(fake.verifyArgsForCall)
 }
 
-func (fake *NymSignatureScheme) VerifyCalls(stub func(handlers.IssuerPublicKey, *math.G1, []byte, []byte) error) {
+func (fake *CredRequest) VerifyCalls(stub func([]byte, handlers.IssuerPublicKey, []byte) error) {
 	fake.verifyMutex.Lock()
 	defer fake.verifyMutex.Unlock()
 	fake.VerifyStub = stub
 }
 
-func (fake *NymSignatureScheme) VerifyArgsForCall(i int) (handlers.IssuerPublicKey, *math.G1, []byte, []byte) {
+func (fake *CredRequest) VerifyArgsForCall(i int) ([]byte, handlers.IssuerPublicKey, []byte) {
 	fake.verifyMutex.RLock()
 	defer fake.verifyMutex.RUnlock()
 	argsForCall := fake.verifyArgsForCall[i]
-	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3, argsForCall.arg4
+	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3
 }
 
-func (fake *NymSignatureScheme) VerifyReturns(result1 error) {
+func (fake *CredRequest) VerifyReturns(result1 error) {
 	fake.verifyMutex.Lock()
 	defer fake.verifyMutex.Unlock()
 	fake.VerifyStub = nil
@@ -177,7 +171,7 @@ func (fake *NymSignatureScheme) VerifyReturns(result1 error) {
 	}{result1}
 }
 
-func (fake *NymSignatureScheme) VerifyReturnsOnCall(i int, result1 error) {
+func (fake *CredRequest) VerifyReturnsOnCall(i int, result1 error) {
 	fake.verifyMutex.Lock()
 	defer fake.verifyMutex.Unlock()
 	fake.VerifyStub = nil
@@ -191,7 +185,7 @@ func (fake *NymSignatureScheme) VerifyReturnsOnCall(i int, result1 error) {
 	}{result1}
 }
 
-func (fake *NymSignatureScheme) Invocations() map[string][][]interface{} {
+func (fake *CredRequest) Invocations() map[string][][]interface{} {
 	fake.invocationsMutex.RLock()
 	defer fake.invocationsMutex.RUnlock()
 	fake.signMutex.RLock()
@@ -205,7 +199,7 @@ func (fake *NymSignatureScheme) Invocations() map[string][][]interface{} {
 	return copiedInvocations
 }
 
-func (fake *NymSignatureScheme) recordInvocation(key string, args []interface{}) {
+func (fake *CredRequest) recordInvocation(key string, args []interface{}) {
 	fake.invocationsMutex.Lock()
 	defer fake.invocationsMutex.Unlock()
 	if fake.invocations == nil {
@@ -217,4 +211,4 @@ func (fake *NymSignatureScheme) recordInvocation(key string, args []interface{})
 	fake.invocations[key] = append(fake.invocations[key], args)
 }
 
-var _ handlers.NymSignatureScheme = new(NymSignatureScheme)
+var _ handlers.CredRequest = new(CredRequest)
