+++ bccsp/handlers/mock/issuer_public_key.go
@@ -4,7 +4,7 @@ package mock
 import (
 	"sync"
 
-	"github.com/IBM/idemix/bccsp/schemes/dlog/handlers"
+	"github.com/IBM/idemix/bccsp/handlers"
 )
 
 type IssuerPublicKey struct {
