+++ bccsp/handlers/mock/nymsignature_scheme.go
@@ -4,17 +4,19 @@ package mock
 import (
 	"sync"
 
-	"github.com/IBM/idemix/bccsp/schemes/dlog/handlers"
+	"github.com/IBM/idemix/bccsp/handlers"
 	math "github.com/IBM/mathlib"
 )
 
-type CredRequest struct {
-	SignStub        func(*math.Zr, handlers.IssuerPublicKey, []byte) ([]byte, error)
+type NymSignatureScheme struct {
+	SignStub        func(*math.Zr, *math.G1, *math.Zr, handlers.IssuerPublicKey, []byte) ([]byte, error)
 	signMutex       sync.RWMutex
 	signArgsForCall []struct {
 		arg1 *math.Zr
-		arg2 handlers.IssuerPublicKey
-		arg3 []byte
+		arg2 *math.G1
+		arg3 *math.Zr
+		arg4 handlers.IssuerPublicKey
+		arg5 []byte
 	}
 	signReturns struct {
 		result1 []byte
@@ -24,12 +26,13 @@ type CredRequest struct {
 		result1 []byte
 		result2 error
 	}
-	VerifyStub        func([]byte, handlers.IssuerPublicKey, []byte) error
+	VerifyStub        func(handlers.IssuerPublicKey, *math.G1, []byte, []byte) error
 	verifyMutex       sync.RWMutex
 	verifyArgsForCall []struct {
-		arg1 []byte
-		arg2 handlers.IssuerPublicKey
+		arg1 handlers.IssuerPublicKey
+		arg2 *math.G1
 		arg3 []byte
+		arg4 []byte
 	}
 	verifyReturns struct {
 		result1 error
@@ -41,25 +44,27 @@ type CredRequest struct {
 	invocationsMutex sync.RWMutex
 }
 
-func (fake *CredRequest) Sign(arg1 *math.Zr, arg2 handlers.IssuerPublicKey, arg3 []byte) ([]byte, error) {
-	var arg3Copy []byte
-	if arg3 != nil {
-		arg3Copy = make([]byte, len(arg3))
-		copy(arg3Copy, arg3)
+func (fake *NymSignatureScheme) Sign(arg1 *math.Zr, arg2 *math.G1, arg3 *math.Zr, arg4 handlers.IssuerPublicKey, arg5 []byte) ([]byte, error) {
+	var arg5Copy []byte
+	if arg5 != nil {
+		arg5Copy = make([]byte, len(arg5))
+		copy(arg5Copy, arg5)
 	}
 	fake.signMutex.Lock()
 	ret, specificReturn := fake.signReturnsOnCall[len(fake.signArgsForCall)]
 	fake.signArgsForCall = append(fake.signArgsForCall, struct {
 		arg1 *math.Zr
-		arg2 handlers.IssuerPublicKey
-		arg3 []byte
-	}{arg1, arg2, arg3Copy})
+		arg2 *math.G1
+		arg3 *math.Zr
+		arg4 handlers.IssuerPublicKey
+		arg5 []byte
+	}{arg1, arg2, arg3, arg4, arg5Copy})
 	stub := fake.SignStub
 	fakeReturns := fake.signReturns
-	fake.recordInvocation("Sign", []interface{}{arg1, arg2, arg3Copy})
+	fake.recordInvocation("Sign", []interface{}{arg1, arg2, arg3, arg4, arg5Copy})
 	fake.signMutex.Unlock()
 	if stub != nil {
-		return stub(arg1, arg2, arg3)
+		return stub(arg1, arg2, arg3, arg4, arg5)
 	}
 	if specificReturn {
 		return ret.result1, ret.result2
@@ -67,26 +72,26 @@ func (fake *CredRequest) Sign(arg1 *math.Zr, arg2 handlers.IssuerPublicKey, arg3
 	return fakeReturns.result1, fakeReturns.result2
 }
 
-func (fake *CredRequest) SignCallCount() int {
+func (fake *NymSignatureScheme) SignCallCount() int {
 	fake.signMutex.RLock()
 	defer fake.signMutex.RUnlock()
 	return len(fake.signArgsForCall)
 }
 
-func (fake *CredRequest) SignCalls(stub func(*math.Zr, handlers.IssuerPublicKey, []byte) ([]byte, error)) {
+func (fake *NymSignatureScheme) SignCalls(stub func(*math.Zr, *math.G1, *math.Zr, handlers.IssuerPublicKey, []byte) ([]byte, error)) {
 	fake.signMutex.Lock()
 	defer fake.signMutex.Unlock()
 	fake.SignStub = stub
 }
 
-func (fake *CredRequest) SignArgsForCall(i int) (*math.Zr, handlers.IssuerPublicKey, []byte) {
+func (fake *NymSignatureScheme) SignArgsForCall(i int) (*math.Zr, *math.G1, *math.Zr, handlers.IssuerPublicKey, []byte) {
 	fake.signMutex.RLock()
 	defer fake.signMutex.RUnlock()
 	argsForCall := fake.signArgsForCall[i]
-	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3
+	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3, argsForCall.arg4, argsForCall.arg5
 }
 
-func (fake *CredRequest) SignReturns(result1 []byte, result2 error) {
+func (fake *NymSignatureScheme) SignReturns(result1 []byte, result2 error) {
 	fake.signMutex.Lock()
 	defer fake.signMutex.Unlock()
 	fake.SignStub = nil
@@ -96,7 +101,7 @@ func (fake *CredRequest) SignReturns(result1 []byte, result2 error) {
 	}{result1, result2}
 }
 
-func (fake *CredRequest) SignReturnsOnCall(i int, result1 []byte, result2 error) {
+func (fake *NymSignatureScheme) SignReturnsOnCall(i int, result1 []byte, result2 error) {
 	fake.signMutex.Lock()
 	defer fake.signMutex.Unlock()
 	fake.SignStub = nil
@@ -112,30 +117,31 @@ func (fake *CredRequest) SignReturnsOnCall(i int, result1 []byte, result2 error)
 	}{result1, result2}
 }
 
-func (fake *CredRequest) Verify(arg1 []byte, arg2 handlers.IssuerPublicKey, arg3 []byte) error {
-	var arg1Copy []byte
-	if arg1 != nil {
-		arg1Copy = make([]byte, len(arg1))
-		copy(arg1Copy, arg1)
-	}
+func (fake *NymSignatureScheme) Verify(arg1 handlers.IssuerPublicKey, arg2 *math.G1, arg3 []byte, arg4 []byte) error {
 	var arg3Copy []byte
 	if arg3 != nil {
 		arg3Copy = make([]byte, len(arg3))
 		copy(arg3Copy, arg3)
 	}
+	var arg4Copy []byte
+	if arg4 != nil {
+		arg4Copy = make([]byte, len(arg4))
+		copy(arg4Copy, arg4)
+	}
 	fake.verifyMutex.Lock()
 	ret, specificReturn := fake.verifyReturnsOnCall[len(fake.verifyArgsForCall)]
 	fake.verifyArgsForCall = append(fake.verifyArgsForCall, struct {
-		arg1 []byte
-		arg2 handlers.IssuerPublicKey
+		arg1 handlers.IssuerPublicKey
+		arg2 *math.G1
 		arg3 []byte
-	}{arg1Copy, arg2, arg3Copy})
+		arg4 []byte
+	}{arg1, arg2, arg3Copy, arg4Copy})
 	stub := fake.VerifyStub
 	fakeReturns := fake.verifyReturns
-	fake.recordInvocation("Verify", []interface{}{arg1Copy, arg2, arg3Copy})
+	fake.recordInvocation("Verify", []interface{}{arg1, arg2, arg3Copy, arg4Copy})
 	fake.verifyMutex.Unlock()
 	if stub != nil {
-		return stub(arg1, arg2, arg3)
+		return stub(arg1, arg2, arg3, arg4)
 	}
 	if specificReturn {
 		return ret.result1
@@ -143,26 +149,26 @@ func (fake *CredRequest) Verify(arg1 []byte, arg2 handlers.IssuerPublicKey, arg3
 	return fakeReturns.result1
 }
 
-func (fake *CredRequest) VerifyCallCount() int {
+func (fake *NymSignatureScheme) VerifyCallCount() int {
 	fake.verifyMutex.RLock()
 	defer fake.verifyMutex.RUnlock()
 	return len(fake.verifyArgsForCall)
 }
 
-func (fake *CredRequest) VerifyCalls(stub func([]byte, handlers.IssuerPublicKey, []byte) error) {
+func (fake *NymSignatureScheme) VerifyCalls(stub func(handlers.IssuerPublicKey, *math.G1, []byte, []byte) error) {
 	fake.verifyMutex.Lock()
 	defer fake.verifyMutex.Unlock()
 	fake.VerifyStub = stub
 }
 
-func (fake *CredRequest) VerifyArgsForCall(i int) ([]byte, handlers.IssuerPublicKey, []byte) {
+func (fake *NymSignatureScheme) VerifyArgsForCall(i int) (handlers.IssuerPublicKey, *math.G1, []byte, []byte) {
 	fake.verifyMutex.RLock()
 	defer fake.verifyMutex.RUnlock()
 	argsForCall := fake.verifyArgsForCall[i]
-	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3
+	return argsForCall.arg1, argsForCall.arg2, argsForCall.arg3, argsForCall.arg4
 }
 
-func (fake *CredRequest) VerifyReturns(result1 error) {
+func (fake *NymSignatureScheme) VerifyReturns(result1 error) {
 	fake.verifyMutex.Lock()
 	defer fake.verifyMutex.Unlock()
 	fake.VerifyStub = nil
@@ -171,7 +177,7 @@ func (fake *CredRequest) VerifyReturns(result1 error) {
 	}{result1}
 }
 
-func (fake *CredRequest) VerifyReturnsOnCall(i int, result1 error) {
+func (fake *NymSignatureScheme) VerifyReturnsOnCall(i int, result1 error) {
 	fake.verifyMutex.Lock()
 	defer fake.verifyMutex.Unlock()
 	fake.VerifyStub = nil
@@ -185,7 +191,7 @@ func (fake *CredRequest) VerifyReturnsOnCall(i int, result1 error) {
 	}{result1}
 }
 
-func (fake *CredRequest) Invocations() map[string][][]interface{} {
+func (fake *NymSignatureScheme) Invocations() map[string][][]interface{} {
 	fake.invocationsMutex.RLock()
 	defer fake.invocationsMutex.RUnlock()
 	fake.signMutex.RLock()
@@ -199,7 +205,7 @@ func (fake *CredRequest) Invocations() map[string][][]interface{} {
 	return copiedInvocations
 }
 
-func (fake *CredRequest) recordInvocation(key string, args []interface{}) {
+func (fake *NymSignatureScheme) recordInvocation(key string, args []interface{}) {
 	fake.invocationsMutex.Lock()
 	defer fake.invocationsMutex.Unlock()
 	if fake.invocations == nil {
@@ -211,4 +217,4 @@ func (fake *CredRequest) recordInvocation(key string, args []interface{}) {
 	fake.invocations[key] = append(fake.invocations[key], args)
 }
 
-var _ handlers.CredRequest = new(CredRequest)
+var _ handlers.NymSignatureScheme = new(NymSignatureScheme)
