+++ bccsp/handlers/revocation_test.go
@@ -14,9 +14,9 @@ import (
 	"encoding/pem"
 	"math/big"
 
+	"github.com/IBM/idemix/bccsp/handlers"
+	"github.com/IBM/idemix/bccsp/handlers/mock"
 	bccsp "github.com/IBM/idemix/bccsp/schemes"
-	"github.com/IBM/idemix/bccsp/schemes/dlog/handlers"
-	"github.com/IBM/idemix/bccsp/schemes/dlog/handlers/mock"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/pkg/errors"
