+++ bccsp/handlers/signer.go
