+++ bccsp/keystore/kvsbased_test.go
@@ -11,10 +11,10 @@ import (
 	"os"
 	"testing"
 
+	"github.com/IBM/idemix/bccsp/handlers"
 	"github.com/IBM/idemix/bccsp/keystore/kvs"
 	bccsp "github.com/IBM/idemix/bccsp/schemes"
 	"github.com/IBM/idemix/bccsp/schemes/dlog/crypto/translator/amcl"
-	"github.com/IBM/idemix/bccsp/schemes/dlog/handlers"
 	math "github.com/IBM/mathlib"
 	"github.com/stretchr/testify/assert"
 )
