+++ bccsp/schemes/dlog/bridge/bridge_test.go
@@ -10,12 +10,12 @@ import (
 	"fmt"
 	"io"
 
+	"github.com/IBM/idemix/bccsp/handlers"
+	"github.com/IBM/idemix/bccsp/handlers/mock"
 	bccsp "github.com/IBM/idemix/bccsp/schemes"
 	"github.com/IBM/idemix/bccsp/schemes/dlog/bridge"
 	idemix "github.com/IBM/idemix/bccsp/schemes/dlog/crypto"
 	"github.com/IBM/idemix/bccsp/schemes/dlog/crypto/translator/amcl"
-	"github.com/IBM/idemix/bccsp/schemes/dlog/handlers"
-	"github.com/IBM/idemix/bccsp/schemes/dlog/handlers/mock"
 	math "github.com/IBM/mathlib"
 	"github.com/golang/protobuf/proto"
 	. "github.com/onsi/ginkgo"
